requirements.txt
arrow==1.3.0
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
inquirerpy==0.3.4
iterfzf==1.0.0.42.0
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
moviepy==1.0.3
numpy==1.26.1
opencv-python==4.8.1.78
pfzy==0.3.4
Pillow==10.0.1
proglog==0.1.10
prompt-toolkit==3.0.39
pyee==11.0.1
Pygments==2.16.1
python-dateutil==2.8.2
python-ffmpeg==2.0.4
requests==2.31.0
rich==13.6.0
six==1.16.0
tqdm==4.66.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.9
